test(replay): split projection tests into focused modules
